// projects/library/alerts/src/lib/tsc/enum/alert.enum.ts
export enum AlertTypeLabelEnum {
    INSTANTANEOUS_LABEL = 'Instantaneous',
    PERIODIC_LABEL = 'Periodic',
    CONSTANT_DATA_LABEL = 'Constant data'
}

export enum AlertTypeEnum {
    INSTANTANEOUS = 'INSTANTANEOUS',
    PERIODIC = 'PERIODIC',
    CONSTANT_DATA = 'CONSTANT_DATA',
    ALL = 'ALL'
}

export enum AlertStatusEnum {
    ACTIVE = 'ACTIVE',
    RESOLVED = 'RESOLVED'
}

export enum AlertStatusLabelEnum {
    ACTIVE_LABEL = 'Active',
    RESOLVED_LABEL = 'Resolved'
}